move pages from common to seperate app
